+++ WSUAutograder/__init__.py
@@ -1,0 +1,2 @@
+from .program import Program, TestCase, TestResult
+from .smartGrader import SmartGrader, Token, TokenType
